lesson_14

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\CategoryRequest;
 use App\Models\Category;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -36,13 +37,19 @@ class CategoryController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(CategoryRequest $request)
     {
-        $file = $request->file('image');
-        $originalName = $file->getClientOriginalName();
-        $path = $file->storeAs('categories', $originalName);
+        $request->validate();
+
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if ($request()->has('image')) {
+            $file = $request->file('image');
+            $originalName = $file->getClientOriginalName();
+            $path = $file->storeAs('categories', $originalName);
+            $params['image'] = $path;
+        }
 
         Category::create($params);
 
@@ -78,15 +85,19 @@ class CategoryController extends Controller
      * @param  \App\Models\Category  $category
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category)
+    public function update(CategoryRequest $request, Category $category)
     {
-        Storage::delete($category->image);
-        
-        $file = $request->file('image');
-        $originalName = $file->getClientOriginalName();
-        $path = $file->storeAs('categories', $originalName);
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            Storage::delete($category->image);
+
+            $file = $request->file('image');
+            $originalName = $file->getClientOriginalName();
+            $path = $file->storeAs('categories', $originalName);
+            $params['image'] = $path;
+        }
 
         $category->update($params);
