+++ app/Http/Controllers/Admin/ProductController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\ProductRequest;
 use App\Models\Category;
 use App\Models\Product;
 use Illuminate\Http\Request;
@@ -38,13 +39,19 @@ class ProductController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(ProductRequest $request)
     {
-        $file = $request->file('image');
-        $originalName = $file->getClientOriginalName();
-        $path = $file->storeAs('products', $originalName);
+        $request->validate();
+
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if ($request()->has('image')) {
+            $file = $request->file('image');
+            $originalName = $file->getClientOriginalName();
+            $path = $file->storeAs('products', $originalName);
+            $params['image'] = $path;
+        }
 
         Product::create($params);
 
@@ -81,15 +88,19 @@ class ProductController extends Controller
      * @param  \App\Product  $product
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductRequest $request, Product $product)
     {
-        Storage::delete($product->image);
-        
-        $file = $request->file('image');
-        $originalName = $file->getClientOriginalName();
-        $path = $file->storeAs('products', $originalName);
         $params = $request->all();
-        $params['image'] = $path;
+        unset($params['image']);
+
+        if ($request->has('image')) {
+            Storage::delete($product->image);
+
+            $file = $request->file('image');
+            $originalName = $file->getClientOriginalName();
+            $path = $file->storeAs('products', $originalName);
+            $params['image'] = $path;
+        }
 
         $product->update($params);
         return redirect()->route('products.index');
